ingest.py: improve split session data detection; more fixes needed.
Using the existing model of checking for existing sessions was incorrectly
checking n>1, implying that n==1 could work, so the resuling n could be 2
before subsequent files were skipped. This is now corrected so that any session
for that date triggers the skip.

However, the logic to handle duplicate files for a day (rig crash)
is still only triggered on the 2nd file ingested, implying that the 1st
partial session could be loaded into the database. A correct fix will
query the RigDataFile table before proceeding rather than the already-ingested
sessions table and is forthcoming.