ingest.py: properly fix duplicate session file logic to query RigDataFile
Fix duplicate session file logic to query RigDataFile -
Previous checks were checking against number of already ingested files for
the date instead of which files were known for the date which the first split
file to be ingested and the remainder to be skipped.

Now, we query the known files for the date so no partial sessions should exist.