ingest.py: RigDataFile sort file subpaths before insertion
This is done to aid the generation of the sequential Session ID's
since the files will be in ascending order according to the date number.

Also of note, the Rig()/RigDataPath() key source should be sorted according
to experimental seqence if these are in stages to ensure proper seqence;
if they are not, more complicated sorting should be done as part of ingest
to keep the generated session ID's in the right sequence or analysis
should further sort by dates, etc.

These things are basically side effects of making the session id synthetic;
future use of native session id's or alternately redesigning around a
non-synthetic value such as session-on-a-day would remove the need for these
workarounds.