ingest.py: ingest rework 1/N: add SessionDiscovery table
New ingest architecture will have script to discover 'known sessions'
e.g. water restrictions + dates found in behavior ingest paths;
this will be used as a key source for behavior and ephys ingest based on
the the presence of corresponding files in the data directories.

The 'SessionDiscovery' table will be the table used to track discovered
sessions.

This is cleaner than having to rely on various undefined behavior such
as insertion seqence in tables, but still allows enough control to ensure
that the synthetic session ID's are properly ordered, but still allows
the 'core' schema to be independent of the ingest method to allow flexibility
in reuse or later adjustments to the ingest mechanism.