map-ephys: initial globus api utilities and unit test
Add pipeline/globus.py which defines a GlobusStorageManager class which provides
a simple set of methods for interacting with globus storage.

The tests/test_globus.py contains unit tests for this class.

While the actual transfer methods and tests are implemented, some
reconfiguration is needed to verify functionality of these methods since using a
personal endpoint does not provide the capability needed for testing, namely,
that files cannot be transferred from/to 2 personal endpoints not tied to
'globus plus' accounts.

Future reconfiguration to a test area in the MAP globus storage area will allow
the functionality/tests to be exercized/corrected.