pipeline/psth.py: initial UnitGroupPsth implementation
Initial implementation of UnitGroupPsth calculation.

Status:

  - Single Dataset computation appears OK via cursory review.
  - Multi set / Plotting/Data review TBD.

Required reworking SelectivityCriteria to allow nullable values in
the 'upper 1/2' of data; these values are used for 'loose' restrictions over
the SelectivityCriteria set which are expanded to the related 'precise'
criteria for Unit selection / computation. For example the value:

  {'sample_selectivity': 1, ... = None}

as used in UnitGroupCondition will be expanded to all SelectivityCriteria
having 1 as a value for 'sample_selectivity', which is then used to query
for matching Units in the UnitGroupPsth calculation. The 'null' values
are *not* produced by the Selectivity calculation, and will have all values
filled. This could potentially be clearer if implemented in a second table;
however this was not done at this time since the two use cases neatly fit
into the two halves of the dataset (1st 1/2: all filled, 2nd 1/2 'nullable').