pipeline/ingest/behavior.py: remove file merge support
background:

  - previous behavior auto-merged >1 files having the same date in the file
    name assuming these were caused by rig crashes
  - new session abstraction allows multiple sessions per day
  - this means auto-merge and determining multi session are undecidable from
    file name alone
  - rig crash problem has been fixed and so file merge is not needed

So, to proceed with multi session per day, remove merge support.

If files need to be merged going forward, this will be done outside
of ingestion.

A secondary implication is that previously, second-run ingest was
determined from file session date; this mechanism breaks for multi-session
case, so we now filter the file traversal for already ingested files
and skip these since session ID is non-deterministic/synthetic ID and
can't be matched from file date alone.