behavior ingest: add mock 'reserve_jobs' to populate to assist devel/debug
implement a fake 'reserve_jobs' in the behavior ingest populate functions
which will emulate the no-exit-on-error behavior of true parallel populate.

this can be set to allow ingest routines to continue to completion to assist
in debugging. If parallel populate is ever implemented, it will retain this
sense.

enable by default for shell 'ingest_foraging_behavior' but not original
'ingest_behavior' - ingest_behavior function shouldn't error as is;
if it does, this means a code bug.